implement helper methods for list table columns [touch: 5564]
Note:  Promotions is currently broken now because the promotion models haven't been added to the addon yet and the core branch I'm testing against does not have models.  So I can't verify code works anyways.

Even after adding the models things will likely be broken for the promotions list table for a bit simply because I'll have to wire in the new methods.  At least by putting them in the code, the errors thrown will make it easier for me to remember which helper methods have to be wired up.