add filters into events list table so if certain request vars are present we filter accordingly [touch: 5866]
This specifically applies to when a promotion applies to multiple events and is linked to the event list table.  We include the event ids so that the event list table is filtered by that promotion.  The PRO_ID is also included so that a title explaining this view linked back to the promotion is displayed as well.