fix wayward click
We were attempting to trigger a click on the SPCO Next Reg Step  button, but were doing so by looking for the first '.spco-next-step-btn' in the  SPCO.main_container, but that was clicking on the button from the previous reg step that was now hidden. Needed to switch to finding the closest() form instead. [touch:11312]